requirements.txt
# Core ML
open_clip_torch>=2.20.0
torch>=2.0.0
torchvision>=0.15.0

# Graph & computation
networkx>=3.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0

# Image handling
Pillow>=9.0.0

# Visualization
matplotlib>=3.7.0

# NLP
spacy>=3.6.0

# Config & utilities
pyyaml>=6.0
requests>=2.28.0
tqdm>=4.65.0

# ROS2 deps (installed via rosdep, not pip):
#   rclpy
#   nav2_msgs
#   sensor_msgs
#   geometry_msgs
#   cv_bridge
#   nav_msgs

# Optional: OpenAI API fallback
# openai>=1.0.0
